remove redundant scss & constants, move stuff from view to component root
