+++ src/react/components/FeatureEditor/DrawingHelper.js
