+++ src/react/components/FeatureEditor/EditConfirmation.jsx
