+++ src/react/components/FeatureEditor/ErrorPanel.jsx
