+++ src/react/components/FeatureEditor/FeatureEditorPanel.jsx
