+++ src/react/components/FeatureEditor/FeatureEditorPanelHandler.js
