+++ src/react/components/FeatureEditor/FeaturePanel.jsx
