+++ src/react/components/FeatureEditor/GeoJSONPanel.jsx
