+++ src/react/components/FeatureEditor/GeometryPanel.jsx
