+++ src/react/components/FeatureEditor/Helper.js
