+++ src/react/components/FeatureEditor/InfoPanel.jsx
