+++ src/react/components/FeatureEditor/index.js
@@ -1,1 +1,1 @@
-export { FeatureEditorPanel } from './view/FeatureEditorPanel';
+export { FeatureEditorPanel } from './FeatureEditorPanel';
