--- src/react/components/FeatureEditor/resources/scss/style.scss
@@ -1,17 +1,0 @@
-#contentMap {
-  &.mapFeatureEditorMode > {
-    #mapdiv, div.row-fluid {
-      position: relative;
-      left: 212px;
-    }
-  }
-}
-
-/* Other */
-#delete-dialog {
-  margin-bottom: 30px;
-}
-
-.field-invalid {
-  border: 1px solid red !important;
-}
